requirements.txt
# AI-Assisted Café Rostering System
# Requirements for v2 Architecture

# Core dependencies
pandas>=2.0.0
pyyaml>=6.0
ortools>=9.8.0  # For CP-SAT constraint programming solver

# Firebase / Firestore
google-cloud-firestore>=2.11.0
google-auth>=2.17.0

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0

# Code quality
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0